requirements.txt
Django==3.2.8
django-multiselectfield==0.1.12
django-import-export==2.6.1
environs==9.3.4
django-phonenumber-field==5.2.0
phonenumbers==8.12.36